+++ Editor/src/com/kotcrab/vis/editor/proxy/AudioProxy.java
@@ -44,7 +44,7 @@ public class AudioProxy extends EntityProxy {
 
 	@Override
 	protected void reloadAccessors () {
-		transform = getEntity().getComponent(Transform.class);
+		transform = getComponent(Transform.class);
 		enableBasicProperties(transform, accessor, accessor);
 	}