+++ Editor/src/com/kotcrab/vis/editor/proxy/PointProxy.java
@@ -41,7 +41,7 @@ public class PointProxy extends EntityProxy {
 
 	@Override
 	protected void reloadAccessors () {
-		transform = getEntity().getComponent(Transform.class);
+		transform = getComponent(Transform.class);
 		enableBasicProperties(transform, accessor, accessor);
 	}